RBAC-system/requirements.txt
asgiref==3.8.1
captcha==0.6.0
certifi==2024.8.30
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
distlib==0.3.8
Django==5.1.1
django-cors-headers==4.4.0
django-redis==5.4.0
django-rest-swagger==2.2.0
djangorestframework==3.15.2
djangorestframework-jwt==1.11.0
filelock==3.15.4
idna==3.10
itypes==1.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mysqlclient==2.2.4
openapi-codec==1.3.2
pillow==10.4.0
platformdirs==4.2.2
PyJWT==1.7.1
redis==5.0.8
requests==2.32.3
simplejson==3.19.3
sqlparse==0.5.1
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.3
virtualenv==20.26.3
